Compile debug augmentation differently.
Uses
  [(...befores, the_call()), ...afters][0]
as JavaScript instead of old
  (...befores, the_call()); ...afters;

Fixes edge case:
super- and/or send-index- augmented
block return turning non-local via inlining.

Failed by having semicolon
in expression context of
  throw $early=[...compiled send...].

Could be solved by finding it in ast preinliner
and set it to alias, but the case is hard
to search for precisely at the ast level,
it is only happening in few cases in IR level.