kernel: Fix nasty error.
st.seamless implementation broken the invariant:
"when there was no context on entry to ST,
there should be no context when leaving ST".

This meant, error handling was "switched off" whenever there was error
in a promises' then: or catch: block (as promises' handlers were
called using st.seamless - there's a good reason for that -
there may be (and should be) .catch somewhere down the chain).

st.seamless is fixed to return thisContext to null on leave
when it was null on entry.